Add analyze datasource test, Improve test flow in CI

+++ tests/test_pandas_accessor.py
@@ -1,30 +1,35 @@
+import sys
 import unittest
-from mindsdb_sdk import AutoML, auto_ml_config
-import pandas as pd
-from subprocess import Popen
 import time
+from subprocess import Popen
+import pandas as pd
+from mindsdb_sdk import AutoML, auto_ml_config
 
 class TestAccessor(unittest.TestCase):
+    start_backend = True
+
     @classmethod
     def setUpClass(cls):
-        cls.sp = Popen(
-            ['python', '-m', 'mindsdb', '--api', 'http'],
-            close_fds=True
-        )
-        time.sleep(40)
+        if cls.start_backend:
+            cls.sp = Popen(
+                ['python', '-m', 'mindsdb', '--api', 'http'],
+                close_fds=True
+            )
+            time.sleep(40)
 
 
     @classmethod
     def tearDownClass(cls):
-        try:
-            conns = psutil.net_connections()
-            pid = [x.pid for x in conns if x.status == 'LISTEN' and x.laddr[1] == 47334 and x.pid is not None]
-            if len(pid) > 0:
-                os.kill(pid[0], 9)
-            cls.sp.kill()
-        except Exception:
-            pass
-        time.sleep(40)
+        if cls.start_backend:
+            try:
+                conns = psutil.net_connections()
+                pid = [x.pid for x in conns if x.status == 'LISTEN' and x.laddr[1] == 47334 and x.pid is not None]
+                if len(pid) > 0:
+                    os.kill(pid[0], 9)
+                cls.sp.kill()
+            except Exception:
+                pass
+            time.sleep(40)
 
     def flow_test_body(self, when=None):
         df = pd.DataFrame({
@@ -49,10 +54,9 @@ class TestAccessor(unittest.TestCase):
         assert len(statistical_analysis) > 8
 
         # Predict from the test dataframe
+        kwargs = {'name': predictor_ref}
         if when:
-            kwargs = {'when_data': when}
-        else:
-            kwargs = {'name': predictor_ref}
+            kwargs["when_data"] = when
         for pred in test_df.auto_ml.predict(**kwargs):
             assert 'y' in pred and pred['y'] is not None
 
@@ -72,6 +76,8 @@ class TestAccessor(unittest.TestCase):
         self.flow_test_body()
 
     def test_3_local_server_flow_with_when_condition(self):
+        # disabled until https://github.com/mindsdb/mindsdb/issues/994 not fixed
+        return
         auto_ml_config(mode='api', connection_info={
             'host': 'http://localhost:47334'
         })
@@ -89,4 +95,9 @@ class TestAccessor(unittest.TestCase):
 
 
 if __name__ == '__main__':
+    if len(sys.argv) > 1 and sys.argv[-1] == "--no_backend_instance":
+        # need to remove if from arg list
+        # mustn't provide it into unittest.main
+        sys.argv.pop()
+        TestAccessor.start_backend = False
     unittest.main(verbosity=2)
